refactor(portal): extract `provideApollo` function
